Redis package had a name conflict with d3a_core.redis package on d3a. Renamed the d3a package to redis_connections.
