+++ src/d3a/d3a_core/d3a_jobs.py
@@ -104,7 +104,7 @@ def start(scenario, settings, events):
                        kwargs=kwargs)
     except Exception:
         import traceback
-        from d3a.d3a_core.redis.redis_communication import publish_job_error_output
+        from d3a.d3a_core.redis_connections.redis_communication import publish_job_error_output
         publish_job_error_output(job.id, traceback.format_exc())
         logging.getLogger().error(f"Error on jobId {job.id}: {traceback.format_exc()}")
