+++ src/d3a/models/area/event_dispatcher.py
@@ -34,7 +34,7 @@ from d3a.models.area.redis_dispatcher.market_event_dispatcher import AreaRedisMa
 from d3a.models.area.redis_dispatcher.area_event_dispatcher import RedisAreaEventDispatcher
 from d3a.models.area.redis_dispatcher.market_notify_event_subscriber \
     import MarketNotifyEventSubscriber
-from d3a.d3a_core.redis.redis_area_market_communicator import RedisCommunicator
+from d3a.d3a_core.redis_connections.redis_area_market_communicator import RedisCommunicator
 
 
 log = getLogger(__name__)
